Teuchos: ETI MpiComm and FancyOStream
